Add suport for local base images

+++ install-arch-linux-rpi-zero-w.sh
@@ -5,33 +5,43 @@ if [ "$(id -u)" == "0" ]; then
 fi
 
 dev=$1
-cd $(mktemp -d)
+tmp=$(mktemp -d)
 
-function umountboot {
-    sudo umount boot || true
-    sudo umount root || true
+function cleanup {
+    sudo umount $tmp/boot || true
+    sudo umount $tmp/root || true
 }
 
-# RPi1/Zero (armv6h):
-archlinux=/tmp/ArchLinuxARM-rpi-latest.tar.gz
-url=http://os.archlinuxarm.org/os/ArchLinuxARM-rpi-latest.tar.gz
+if [ $# -lt 2 ] ; then
+  # RPi1/Zero (armv6h):
+  archlinux=/tmp/ArchLinuxARM-rpi-latest.tar.gz
+  url=http://os.archlinuxarm.org/os/ArchLinuxARM-rpi-latest.tar.gz
 
-# RPi2 (armv7h):
-# archlinux=/tmp/ArchLinuxARM-rpi-2-latest.tar.gz
-# url=http://os.archlinuxarm.org/os/ArchLinuxARM-rpi-2-latest.tar.gz
+  # RPi2 (armv7h):
+  # archlinux=/tmp/ArchLinuxARM-rpi-2-latest.tar.gz
+  # url=http://os.archlinuxarm.org/os/ArchLinuxARM-rpi-2-latest.tar.gz
 
-curl -L -o $archlinux -z $archlinux $url
+  curl -L -o $archlinux -z $archlinux $url
+else
+  archlinux=$2
+fi
+
+if ! [ -f $archlinux ] ; then
+  echo "File not found: $archlinux"
+  exit 1
+fi
+
 sudo parted -s $dev mklabel msdos
 sudo parted -s $dev mkpart primary fat32 1 128
 sudo parted -s $dev mkpart primary ext4 128 -- -1
 sudo mkfs.vfat ${dev}1
 sudo mkfs.ext4 -F ${dev}2
-mkdir -p boot
-sudo mount ${dev}1 boot
-trap umountboot EXIT
-mkdir -p root
-sudo mount ${dev}2 root
+mkdir -p $tmp/boot
+sudo mount ${dev}1 $tmp/boot
+trap cleanup EXIT
+mkdir -p $tmp/root
+sudo mount ${dev}2 $tmp/root
 
-sudo bsdtar -xpf $archlinux -C root
+sudo bsdtar -xpf $archlinux -C $tmp/root
+sudo mv $tmp/root/boot/* $tmp/boot
 sudo sync
-sudo mv root/boot/* boot
